Convenience methods for PathObjectHierarchy
Add more convenience methods for getting objects from a `PathObjectHierarchy` without needing to worry about the specific Java class used for object types.

This includes:
```
getAllObjectsForROI(ROI roi)
getAnnotationsForROI(ROI roi)
getTilesForROI(ROI roi)
getCellsForROI(ROI roi)
getAllDetectionsForROI(ROI roi)
```

Also add `ROIs.createEllipseROI(ImageRegion region)` for convenience, similar to how a rectangle ROI can be made.

Along the way, create a new class for hierarchy tests - since the older ones are hard to read and maintain.
